src/saaga_mcp_servers/agents/ambient/pyproject.toml
[project]
name = "ambient"
version = "0.1.0"
description = "Lightweight Jira monitoring agent that triggers Claude code execution"
readme = "README.md"
authors = [
    { name = "Andrew Oseen", email = "amovfx@protonmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "conduit-connect",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
    "fire>=0.7.0",
    "assemblyai[extras]>=0.41.1",
    "langchain[anthropic]>=0.3.19",
    "pydantic-settings>=2.9.1",
]

[project.scripts]
amb = "ambient.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
